Create Tyler Blakemore Reading 01
The first reading was a really good insight to what light is and how it actually behaves. As a Graphic Designer and Photographer, this reading helped me understand light and why it works the way that it does. One particular section that I thought was really interesting was the section 2.1 on reflection. Even at the microscopic levels there can be many different reflection points happening at one time, depending on how uneven the surface is and the angle of light. In section 4.1 we start to geting hue, saturation and brightness and the author broke these down to make it really easy to understand that hue is a spectrum on which colors fall onto, while saturation is the purity of a color and brighness is the intensity of the color.