Alejandro_garcia_reading02.md
After watching the video on Glitch art I already find it very interesting since I too have dabbled in glitch art as well using effects and filters on the app "Picsart" however I am nowhere near as skilled as any of the mentioned artist and can not even edit a video, but I still like the content I created and enjoyed making it. I like that there are meetings for fellow glitch artist where they can meet and if one is ever near me I plan to attend. Although I didn't try it I may attempt to use the glitch art presented in video 1 where he says to drag a jpeg and mess around with the numbers 0 and 1 in the photo he added. 

In video 2 I found it interesting when mentioned that "glitch=unexpected" and that computers do not make mistakes, humans do. That a computer did what was asked but you are the one at fault for using your glitch. Another thing I found interesting is that glitch art does not equal glitch which is how wikipedia describes it. The narrator goes into detail explaining how it's wrong that a "glitch is an unintentional glitch in the system, but glitch art is intentionally done and that an artist can use the aestheticization to have art artifacts and data bending" using the keyword intentions