Create Tyler_Blakemore_Reading2.md
The second reading really inspired my first project. I decided to go with glitch art because I really felt that the style was something I was interested in. The definition that was given “Take a familiar piece of technology and do something unfamiliar” really resonated with me because it was very open-ended. And when the author Nick Briz was going over the different ways and techniques, I felt very inspired because I saw so much opportunity. The other thing about this article and overall glitch are is that it is open-ended. There is no ugly, there just is. Knowing that, I felt very drawn to this style.